fixed typo
